refactor: add shared header with slots
