+++ src/widgets/header-profile/index.ts
@@ -1,0 +1,1 @@
+export { HeaderProfile } from './ui/HeaderProfile'
